refactor: Person inside PageMetadata to a component
